ecommerce/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==4.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.2
razorpay==2.0.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
